addresses comments.
